SideBar: cambio de estilos, correccion en actions, sweet alert implementado, formulario de creacion de casos

+++ client/src/views/Dashboard/AdoptionsDashboard/AdoptionsDashboard.jsx
@@ -1,13 +1,13 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import Swal from 'sweetalert2';
 import { faPlus, faPenToSquare, faTrash } from '@fortawesome/free-solid-svg-icons';
 import { useDispatch, useSelector } from 'react-redux';
 import { useEffect } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link } from 'react-router-dom';
 import { getAdoptions, deleteAdoptionById } from '../../../redux/adoptions/adoptionsActions';
 
 const AdoptionsDashboard = () => {
     const dispatch = useDispatch();
-    const navigate = useNavigate();
     const { adoptions } = useSelector(state => state.adoptions);
 
     useEffect(() => {
@@ -15,8 +15,25 @@ const AdoptionsDashboard = () => {
     }, [dispatch]);
 
     const handleClick = (id) => {
-        dispatch(deleteAdoptionById(id))
-    }
+        Swal.fire({
+          title: "Are you sure?",
+          text: "You won't be able to revert this!",
+          icon: "warning",
+          showCancelButton: true,
+          confirmButtonColor: "#3085d6",
+          cancelButtonColor: "#d33",
+          confirmButtonText: "Yes, delete it!"
+        }).then((result) => {
+          if (result.isConfirmed) {
+            Swal.fire({
+              title: "Deleted!",
+              text: "Your file has been deleted.",
+              icon: "success"
+            });
+            dispatch(deleteAdoptionById(id));
+          }
+        });
+      }
 
     return (
         <section className="flex justify-center sm:ml-64">
